put configLoader to npm package
