+++ lib/generators/react_on_rails/dev_tests_generator.rb
@@ -45,28 +45,14 @@ module ReactOnRails
         File.open(package_json, "w+") { |f| f.puts new_contents }
       end
 
-      # Since we use the local react-on-rails package, we need an alias for it
-      def add_alias_for_react_on_rails
-        config = File.join(destination_root, "client", "webpack.config.js")
-        old_contents = File.read(config)
-        new_contents = old_contents.gsub(
-          /extensions: \['\.js', '\.jsx'\],/,
-          'extensions: [\'.js\', \'.jsx\'],
-            alias: {
-              \'react-on-rails\': resolve(__dirname, \'..\', \'..\', \'..\', \'..\'),
-            },'
-        )
-        File.open(config, "w+") { |f| f.puts new_contents }
-      end
-
-      def change_webpack_client_base_config_to_include_fallback
-        sentinel = /^\s\s},\n\s\splugins: \[\n/
-        config = File.join(destination_root, "client", "webpack.config.js")
-        old_contents = File.read(config)
-        new_contents = "const path = require('path');\n" +
-                       old_contents.gsub(sentinel, FALLBACK_OPTION_FOR_NODE_MODULES)
-        File.open(config, "w+") { |f| f.puts new_contents }
-      end
+      # def change_webpack_client_base_config_to_include_fallback
+      #   sentinel = /^\s\s},\n\s\splugins: \[\n/
+      #   config = File.join(destination_root, "client", "webpack.config.js")
+      #   old_contents = File.read(config)
+      #   new_contents = "const { resolve } = require('path');\n" +
+      #                  old_contents.gsub(sentinel, FALLBACK_OPTION_FOR_NODE_MODULES)
+      #   File.open(config, "w+") { |f| f.puts new_contents }
+      # end
 
       def add_test_related_gems_to_gemfile
         gem("rspec-rails", group: :test)
