+++ lib/generators/react_on_rails/templates/base/base/client/webpack.config.js
@@ -4,8 +4,10 @@
 
 const webpack = require('webpack');
 const { resolve } = require('path');
-const ManifestPlugin = require('webpack-manifest-plugin'); // eslint-disable-line import/no-unresolved
-const { env, paths, publicPath } = require('./webpackConfigLoader.js');
+const ManifestPlugin = require('webpack-manifest-plugin');
+const webpackConfigLoader = require('react-on-rails').default.webpackConfigLoader; // eslint-disable-line import/no-unresolved
+
+const { env, paths, publicPath } = webpackConfigLoader(resolve('..', 'config', 'webpack'));
 
 const devBuild = env !== 'production';
