+++ node_package/src/ReactOnRails.js
@@ -8,6 +8,7 @@ import serverRenderReactComponent from './serverRenderReactComponent';
 import buildConsoleReplay from './buildConsoleReplay';
 import createReactElement from './createReactElement';
 import Authenticity from './Authenticity';
+import webpackConfigLoader from './webpackConfigLoader';
 import context from './context';
 
 const ctx = context();
@@ -221,6 +222,10 @@ ctx.ReactOnRails = {
 
   resetOptions() {
     this.options = Object.assign({}, DEFAULT_OPTIONS);
+  },
+
+  webpackConfigLoader(configPath) {
+    return webpackConfigLoader.configLoader(configPath);
   },
 };
