+++ spec/dummy/client/webpack.server.rails.build.config.js
@@ -8,7 +8,8 @@ const webpackCommon = require('./webpack.common');
 const { assetLoaderRules } = webpackCommon;
 
 const ManifestPlugin = require('webpack-manifest-plugin');
-const { paths, publicPath } = require('./webpackConfigLoader.js');
+const webpackConfigLoader = require('react-on-rails').default.webpackConfigLoader;
+const { env, paths, publicPath } = webpackConfigLoader(resolve('..', 'config', 'webpack'));
 const manifestPath = resolve('..', paths.output, paths.assets, paths.manifest);
 
 const devBuild = process.env.NODE_ENV !== 'production';
@@ -36,7 +37,7 @@ module.exports = {
     extensions: ['.js', '.jsx'],
     alias: {
       images: join(process.cwd(), 'app', 'assets', 'images'),
-      'react-on-rails': resolve(__dirname, '..', '..', '..'),
+      // 'react-on-rails': resolve(__dirname, '..', '..', '..'),
     },
   },
   plugins: [
